pyproject.toml
[project]
name = "personal-ai-employee"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-api-python-client>=2.190.0",
    "google-auth>=2.48.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "mcp>=1.26.0",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "tweepy>=4.16.0",
    "watchdog>=6.0.0",
]
